test(logstreams): rewrite LogStorageAppenderTest
Rewrites the `LogStorageAppenderTest`, moving it back to the logstreams
module, and updating it for JUnit 5. The inconsistency detection test is
rewritten for the new gap check implemented in the previous commit.
